Improve rendering of pull request template in GitHub
All that is changed is that list items are no longer broken up into
multiple lines and a markdownlint ignore directive is added for one
particularly long markdown line.  This improves rendering in GitHub
because the list items no longer appear artificially shortened and
instead can be rendered as markdown sees fit.